move asm common functions to lib
