+++ src/backends/kimchi/mod.rs
@@ -12,7 +12,7 @@ use itertools::{izip, Itertools};
 use kimchi::circuits::polynomials::generic::{GENERIC_COEFFS, GENERIC_REGISTERS};
 
 use crate::{
-    backends::kimchi::asm::{display_source, parse_coeffs},
+    backends::kimchi::asm::parse_coeffs,
     circuit_writer::{
         writer::{AnnotatedCell, Cell, PendingGate},
         DebugInfo, Gate, GateKind, Wiring,
@@ -27,7 +27,7 @@ use crate::{
 
 use ark_ff::{One, Zero};
 
-use self::asm::{extract_vars_from_coeffs, title, OrderedHashSet};
+use self::asm::{extract_vars_from_coeffs, OrderedHashSet};
 
 /// We use the scalar field of Vesta as our circuit field.
 pub type VestaField = kimchi::mina_curves::pasta::Fp;
@@ -467,7 +467,7 @@ impl Backend for KimchiVesta {
         }
 
         if debug && !vars.is_empty() {
-            title(&mut res, "VARS");
+            crate::helpers::title(&mut res, "VARS");
         }
 
         for (idx, var) in vars.iter().enumerate() {
@@ -476,7 +476,7 @@ impl Backend for KimchiVesta {
 
         // gates
         if debug {
-            title(&mut res, "GATES");
+            crate::helpers::title(&mut res, "GATES");
         }
 
         for (row, (Gate { typ, coeffs }, debug_info)) in
@@ -506,7 +506,7 @@ impl Backend for KimchiVesta {
 
             if debug {
                 // source
-                display_source(&mut res, sources, &[debug_info.clone()]);
+                crate::helpers::display_source(&mut res, sources, &[debug_info.clone()]);
 
                 // note
                 res.push_str("    ▲\n");
@@ -519,7 +519,7 @@ impl Backend for KimchiVesta {
 
         // wiring
         if debug {
-            title(&mut res, "WIRING");
+            crate::helpers::title(&mut res, "WIRING");
         }
 
         let mut cycles: Vec<_> = self
@@ -544,7 +544,7 @@ impl Backend for KimchiVesta {
                 .unzip();
 
             if debug {
-                display_source(&mut res, sources, &debug_infos);
+                crate::helpers::display_source(&mut res, sources, &debug_infos);
             }
 
             let s = cells.iter().map(|cell| format!("{cell}")).join(" -> ");
